update tests and put_checksum_algorithm
