+++ tests/s3_data_plane_tests.c
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
+#include "aws/s3/private/s3_checksums.h"
 #include "aws/s3/private/s3_client_impl.h"
 #include "aws/s3/private/s3_meta_request_impl.h"
 #include "aws/s3/private/s3_util.h"
@@ -2201,6 +2202,40 @@ static int s_test_s3_round_trip_default_get(struct aws_allocator *allocator, voi
     return 0;
 }
 
+int s_s3_validate_headers_checksum_set(
+    struct aws_s3_meta_request *meta_request,
+    const struct aws_http_headers *headers,
+    int response_status,
+    void *user_data) {
+    (void)response_status;
+    (void)headers;
+    struct aws_s3_meta_request_test_results *meta_request_test_results =
+        (struct aws_s3_meta_request_test_results *)user_data;
+    ASSERT_NOT_NULL(meta_request->meta_request_level_running_response_sum);
+    ASSERT_INT_EQUALS(
+        meta_request->meta_request_level_running_response_sum->algorithm, meta_request_test_results->algorithm);
+    return AWS_OP_SUCCESS;
+}
+
+void s_s3_test_validate_checksum(
+    struct aws_s3_meta_request *meta_request,
+    const struct aws_s3_meta_request_result *result,
+    void *user_data) {
+    (void)meta_request;
+    (void)user_data;
+    AWS_FATAL_ASSERT(result->did_validate);
+    AWS_FATAL_ASSERT(result->error_code == AWS_OP_SUCCESS);
+}
+void s_s3_test_no_validate_checksum(
+    struct aws_s3_meta_request *meta_request,
+    const struct aws_s3_meta_request_result *result,
+    void *user_data) {
+    (void)meta_request;
+    (void)user_data;
+    AWS_FATAL_ASSERT(!result->did_validate);
+    AWS_FATAL_ASSERT(result->error_code == AWS_OP_SUCCESS);
+}
+
 AWS_TEST_CASE(test_s3_round_trip_default_get_fc, s_test_s3_round_trip_default_get_fc)
 static int s_test_s3_round_trip_default_get_fc(struct aws_allocator *allocator, void *ctx) {
     (void)ctx;
@@ -2254,8 +2289,8 @@ static int s_test_s3_round_trip_default_get_fc(struct aws_allocator *allocator, 
             {
                 .mode = AWS_S3_TESTER_DEFAULT_TYPE_MODE_GET,
             },
-        .finish_callback = s3_test_validate_checksum,
-        .headers_callback = s3_validate_headers_checksum_set,
+        .finish_callback = s_s3_test_validate_checksum,
+        .headers_callback = s_s3_validate_headers_checksum_set,
     };
 
     ASSERT_SUCCESS(aws_s3_tester_send_meta_request_with_options(&tester, &get_options, NULL));
@@ -2316,8 +2351,8 @@ static int s_test_s3_round_trip_fc(struct aws_allocator *allocator, void *ctx) {
             {
                 .object_path = object_path,
             },
-        .finish_callback = s3_test_validate_checksum,
-        .headers_callback = s3_validate_headers_checksum_set,
+        .finish_callback = s_s3_test_validate_checksum,
+        .headers_callback = s_s3_validate_headers_checksum_set,
     };
 
     ASSERT_SUCCESS(aws_s3_tester_send_meta_request_with_options(&tester, &get_options, NULL));
@@ -2357,6 +2392,7 @@ static int s_test_s3_round_trip_mpu_fc(struct aws_allocator *allocator, void *ct
                 .object_size_mb = 10,
                 .object_path_override = object_path,
             },
+        .finish_callback = s_s3_test_no_validate_checksum,
     };
 
     ASSERT_SUCCESS(aws_s3_tester_send_meta_request_with_options(&tester, &put_options, NULL));
